Quick fix for last schema update.
There was an issue with the last commit's core data schema update.
It didn't cause a crash but it did cause new entries to no longer
save properly. For now, since we're so early in the process, rather
than do data migration, this patch simply changes the data store
file name, which causes a new DB file to be created with the new
schema. This in effect resets the data, but fixes the no-new-data
issue.

Change-Id: I5d139e1743eeb13551bdda2d87a8fce0e4381df4