Table of contents ux refinement.
Fix for image taps causing mobile web site to load.

Edit icons now take action on touch up instead of
touch down so attempting to scroll and having the
scroll touch down on an edit icon no longer jumps
to edit screen.

Table of contents images seamlessly update if
downloaded while table of contents is onscreen.

Fixed bug with article scroll position not being
remembered if it was scrolled all the way to the top.

Table of contents can be dragged to show and dismiss.

Change-Id: I0cbe74ece7ad3cf61295d3b23fa31f99a62ff4fc