Adding userId for edit event logging
Note each wiki has a separate local user ID for each local user,
so a single global SUL account will need to 'change' IDs
depending on the language we're working with.

Solving that by appending a 'meta=userinfo' query in when loading
section text into the editor, and passing that into the EL funnel.

Also normalizing username on login, which is more reliable globally.

Change-Id: Ied5d52cf4da657f06c4b23a5cba13094e228b766