Fix ToC scrolling to sections near end of page
The ToC section tap code was scrolling past the end of the document,
which got reset as soon as we adjusted the document further (say by
zooming in when hiding the ToC).

Work around this by saving the scroll position before we zoom out,
and setting it back again after the animation. There is a brief
flash in the case where we had to move, but it works on iOS 6 and 7
and has less scary bridge code than previous patch attempt. :D

Bug: 67633
Change-Id: Idc6f454bc35fd621b3434734a0629ad750d441d0