Quick hack to work around image cache bugs on iOS 8 Retina displays
WebKit in iOS 8 understands 'srcset' attribute and automatically
loads up the '2x' images on a Retina display, which wasn't happening
on iOS 6 or 7.

This commit tweaks the pre-cached-images parsing on section HTML
to also understand srcset, just enough to grab the 2x images and
insert them into our to-be-cached list.

This gets section images cached again on iOS 8 Retina devices,
and makes the images show up in Recent and Saved Pages as expected.

Behavior is left unchanged on older versions of iOS.

Change-Id: I83043ae33703f7de03a6cd248d969afc13c44b4d