Workaround for iOS 8 beta view dismiss buggage with multiple modals
When dismissing multiple modal popups together, documentation states
that only the topmost one is animated, and any intermediaries simply
disappear without animation.

However in iOS 8 beta 5 it animates the *bottommost* modal, instead
of the topmost. This is very annoying.

As workaround, on 8 we take a snapshot of the topmost modal, copy
thta and its transition settings onto the bottommost, and then
dismiss.

Feels hacky but it seems to work. Sigh.

Change-Id: Ie983a96beaf4355ca70055f09a7d05328cd9347a