Fixes for statusbar being hidden in landscape on iOS 8
Now checks for the actual status bar size and makes sure the
top menus get positioned and sized based on that.

Note there is an oddity where after doing a search or something
the status bar starts showing up in landscape again.... Might
have to tweak some of the other code about things wanting/not
wanting status bar to be adjusted.

Change-Id: Icb353a36c646a95300e9904178d88399ccacde92