Clear the webview immediately when loading a new page.
From a UX perspective we want to clear the contents of the current page
as quickly as possible when we get the request for a new one. Setting
the content div innerHTML to an empty string via javascript is the
fastest and most efficient way to do this. I did not implement a
spinner as mentioned in the ticket.

Bug: 70422
Change-Id: Idd0adb963619e2abdaaefce7e4e6e7b38b42b7dd