Support DISPLAYTITLE in article view
Added support for the display title. Limited this to
article view as this seems to be the consistent behavior
across platforms. There was actually a field for this in the
article model, but it was apparently never fully implemented.
Perform nil check when create dictionary in
DownloadSectionsOp.m, and in getHeaderTitle
Section+DisplayHtml.m.

Bug: 63971
Change-Id: I26fe6d723540b67da533e6b916a6a288a4ddc4c9