Make zero-rating dialog and inspection less intrusive
When not in the webview, screens like search were prompting
excessively about zero-rating. Now, these screens outside of the
webview will usually not prompt about zero-rating, unless the
state of zero-rating has gone from On to Off, which is an event
worthy of a pause by the user. There is a chance that the user
may use the search screen and go into zero-rating mode On for
the very first time since installing the app. But that's a pretty
low probability event. It's much more likely that the user will
go into zero-rating mode On from the first time launch screen
or article reading mode. But if the low probability event does
occur, the user will get a prompt. Otherwise, if the user is in
a search screen the second or subsequent time zero-rating goes
into On mode, the user will still have a fixed bar back in the
webview screen showing zero-rating On and will also have the
More... screen to learn more about zero-rating.

In addition to the alert reduction, the app will now only check
for the outbound zero-rating applicability header, "X-CS",
when the URL is a .m.wikipedia.org subdomain. Presently, this is
the only place where such outbound header tagging can occur, so
there is no point initiating a "banner text" fetch on other
domains. This should have the fringe benefit of errant transitions
for On and Off triggering when domains like www.wikidata.org are
accessed by the app. A similar change has been enacted in the
Android app master branch already. Operators zero-rating apps traffic
zero-rate based upon well defined destination IP addresses and
domain names, which include subdomains of wikidata.org. In the
cases where other facets of the request may disqualify otherwise
zero-rated access, backoffice configuration supports the notion
of disabling zero-rating messaging on a given operator's network.

Change-Id: I04e319530af8986081ee9730b3ddd5f7af553230