Add X-WMF-UUID to Wikipedia for iOS requests
For now, the inclusion of the UUID in other places not to be removed.

In practice, it seems the @"ReadingAction" string value is used for
initialization of the UUID in different funnels, so it would seem
that use of the ReadingActionFunnel, which initializes with this
same key, is a safe way to ensure an accurate UUID without exposing
internals. Perhaps ReadingActionFunnel should be renamed to something
else, but that's a separate matter. For now the goal is to ensure the
presence of the accurately set X-WMF-UUID header without causing other
unintended side effects in the app.

It should be possible to query request logs once this code and its
counterpart, in the following change

Ieeb3cc38e8d2f3244ec63d5345d4556e2f4a5df9

, are both deployed. If the values match when expected to match,
then it should be possible to remove the UUID in the other contexts
(e.g., ArticleFether.m).

Change-Id: Ia0ab1691a31b9555721291803481104cb808689c