Don't include article snippet in "Share as text" option
Presently, selecting the "Share as text" option when you share something
and haven't actually highlighted any text will cause the auto-generated
snippet to be shared, which generates shares so big they can't even be shared
to Twitter.

This patch changes "Share as text" when you've got nothing hightlighted to
not attempt to share the default snippet. This should fix that issue for now,
until it's refined further.

The "Share as text" option now no longer will attach the stock image, either.

Additionally, this patch lowers the minimum text selection requirement to two
characters. Finally, it updates the logging "target"s to entered_image and
entered_text, respectively, based on user tap, while also letting the system
perform text truncation with an ellipses even for "Share as text" when
text of sufficient length is not already highlighted.

Bug: T89541
Change-Id: Ic8c749d9207d7e28043ecba896a3c2326dd85a34