TOC performance Bug:T91087 Bug:T91330
Instead of creating the TOC for the current article before it displays,
use the DOM loaded notification and create it after all the web work is complete.
Update the TOC after the spinner is dismissed (instead of before - makes the page feel like it loads faster).
Only update the TOC in real time if visible.
If hidden, update the TOC only after scrolling has ended.
Create an property to hold the highlighted index instead of querying section array

Change-Id: Ib96275e9f7995b278b33cc56cad84da4678d5b29