uncrustify tweaks
This patch is a compromise that indents blocks consistently, as
seen in WMFCodingStyle.m.  The team would prefer Xcode-style block
indentation, but it is still not without issues (see links below).
The current configuration will at least be consistent, and we can
reattempt xcode-style indentation when it has fewer unsupported edge
cases.

https://github.com/bengardner/uncrustify/issues/68
https://github.com/bengardner/uncrustify/issues/109
https://github.com/bengardner/uncrustify/issues/241
https://github.com/JoshuaGross/uncrustify/commit/b4c8b3610166b580f557e0f90f9c091b63830e25

Change-Id: Ic0e8e0a08325d37aa1a59364e25c7e21fbcb41e8