Fix for tap on small images not causing gallery view to appear.
To test, go to "Fish" on enwiki, then go to the "Muscular system"
section. The swim bladder image will not respond to tap. With this
patch it does.

Background:

We don't want to route tiny chrome images, of which there are many,
to our custom data store. The original threshold I had chosen was
100x100, which is too large.

(^For an example of a "tiny chrome image" see the "See also" section
of the "Fish" article.)

The image gallery code assumes that if you are seeing an image, it
has been routed to the data store (how else could you tap it if you
couldn't see it). With the threshold set to 100x100, images that
are large enough to be reasonably expected to be tappable were not
being routed to the data store.

Change-Id: I510b457d86c2b950f5adb1d8698c554eb73aefd6