Don't invoke native bridge if there is selected text
If there is selected text the user is involved with a process that our native code is largely not
invloved in. The property "rangeCount" indicates the number of ranges that are selected. If there
is at least one (the only number that can be selected with human interface on the iphone) then we
know that there is selected text, and the user should complete the flow involved with the selected
text before our native code reacts to touch interactions.

The alternative to this was to pass the rangeCount as a parameter to the native code, but this
would have required an increase of the interface between the javascipt and native code, which
should probably stay as simple as possible.

From a usability standpoint this means that if there is selected text, then a user will first need
to tap to deselect the text, then tap to show the menu again. As a user I believe this is a
predictable interaction and should be fine.

Bug: T91605
Change-Id: I63640ef374034b093b6dafedf8b5c3e35d4a04c3