Only calculate cell height if the length of the text will be "too long" to fit in a normal size cell.
This code is more of a hack…
It makes an assumption of cell width and character width - and is really just a fuzzy metric
to see if we need to kick in the auto layout magic.
There is a decent probability that this will cause some visual glitches if the metric is wrong (e.g. cell overlap)

Since older devices are fixed width and that is the main concern of this patch,
the tradeoff here should be fine.
But it needs some good testing during regression to make sure the thresholds are correct.

If we can't properly address glitches during regression, the best course will be to make cells
fixed height on the older devices and truncate text to fit.

Change-Id: Ib613f1cdd2dbc5fd780b65a1f1443767310fd52d