Change the heading for privacy and terms to 'privacy and terms'
Alters the header in the menu to read 'Privacy and Terms', for a
variety of reasons. The high-level summary is, in ascending
order of importance:
1. The current terminology is idiomatic, and it's probably not
a good idea to expect every user to understand idioms, particularly
US/GB-dominant idioms.
2. The current terminology is patronising; it suggests that
how much data you send us and under what (legally binding, criminally
enforced) terms you contribute to and use our products, and under
what terms we use and reuse and distribute your data, is
impossible-to-understand nonsense you shouldn't bother yourself with.
3. The current terminology is offensive, to me and undoubtedly
to many others: it originated in British attitudes to
African belief structures, and the opinion that they were quaint and
ignorant.

Android doesn't have this heading (it's under a generalised "about")
so only iOS needs patching.

Change-Id: I98d3da0abdf1cbe39b340d73ae2b3287c01911d6