Fastfile updates
Adding 3 things:
Comments for clean and archive… to remove some of the confusion.
Add back a commented out verbose flag so its clear how to enable verbose logging for xcodebuild
Reset the repo on successful exit… this does not happen if we have an error so we can inspect what went wrong.

Change-Id: I449972360bfb408e5d211803adc1b23e38af3467