commit web assets and remove build scripts
This removes node as a build dependency, making setup easier.

Updating web assets can now be done manually using the "web" target in
the Makefile which will:

- Download latests mobile app styles
- Install npm modules
- Run grunt (linting & processing JS code)

Remaining script build phases in the project are for processing the app
icon and copying them into the app's "assets" folder (so the webview can
grab the app icon).

Bug: T94231
Change-Id: Ic535eb9644d66e3632e353544b82494c859d8a7a