Fix for T94534 not working on iPads
Shell script was setting names based on the assets file but file names in the catalog are apparently is not uniform.
The fix is using the files in the app bundle for the source.
This actually simplifies the script quite a bit, so win win!
Additionally had to do some dangling to loop over paths that have spaces

Change-Id: Ic3f909e250b7eddfd115131c4115a7aef09b8c1b