gallery: re-add out of bounds assertion
Fetching indices out of bounds is a programmer error, and should be an
assertion. The real problem causing T95601 was that the call to
"fetchBatchContainingIndex:withNthNeighbor:" was attempting to fetch out
of bounds. This patch fixes that problem and puts the assertion back in
place (to give us the feedback that was originally necessary to find issue)
while preserving fallback behavior when assertions are disabled in an attempt
to allow the app to continue to function (albeit in a semi undefined
state).

The original bug could be reproduced by scrolling any of the 5th last 
images in a gallery.  I reproduced the assertion by doing this when it 
was added, then added a unit test w/ a fix.

Bug: T95601
Change-Id: Ie2bf16998726de73b98a8c8e9e9cb5f60fca8b8c