Fixing icons… again
A few problems to overcome this time:
Archives use symlinks for the app bundle, which can throw monkey wrenches in the process.
Looks like permissions have changed on build folders - can't reliably save to the app bundle during the process. So sidestepping.

New solution:
No longer manipulating the build folder, doing everything in the source directory.
Created a new icon set - source - which will always contain the unmolested icons.
The other icon set will be overwritten at build time with the overlaid icons.
If release build or if imagemagick is not installed, then the icons are just copied
Also removed generated icons from source control

Note: only one icons is copied to the app, the other is in the project only.

Change-Id: I39285429803b72d5caeda738fb82bc3d23406317