Bug T96605 - fixing saves of the history
Another case of throwing exceptions in the data layer.
This patch adds support for NSErrors in place of exceptions.
To do this required changing the save method to use one that produces an error, and
making changes to some method signatures to bubble the errors up.
Also, changed the flow of the saves a bit, instead of passing a save block,
methods just need to convert the object into NSData before writing to disk.

Change-Id: I1e6fb1fad6b93277268a6c7c7df42424c80fa83d