do text sanitation on selected text before sharing
We weren't doing sanitation on selected text before sharing, which was
allowing things like citations to slip through. See added test case for
wmf_shareSnippetFromText and its usage (as opposed to
wmf_shareSnippetFromHTML).

One drawback to this approach is that if a user *partially* selects a
citation "like this.[1" it will still be passed through. IMO we need a
more comprehensive solution for how we select text for sharing (perhaps
using selectors on HTML w/ Parsoid markup?).

Bug: T95476
Change-Id: Id3e0d26c697d409baf6668118a8e695ff96ab4e9