bump fastlane to 1.0, simplify fastfile
This started as cleanup/fix-attempts for T98504, but it's still not
fixed. Need to reach out to fastlane team for help.

In the meantime, and since they'll probably ask us to upgrade anyway, I
bumped us to 1.0 and:

- Remove calls to produce (app records already exist)
- Kept test actions in separate lane, and chain them manually in Jenkins

I basically moved our fastlane folder to a tmp location, re-ran fastlane
init, and added lanes for other configs/app-ids.

Change-Id: Iab9e2f32f1a41f4b7a6b75db47948aa2f3047a9e