Removing needsRefreshed from Article Bug T98907 Follow on refactor for T97781
This was being stored in the model, but was a view layer concern.
Additionally this was not needed "later" needsRefresh is a "now" concern - so
storage was unnecessary.
Replaced with a scoped method variable where it is needed (see:WebVC line:1394).
(aside: conceivably we could have a needsRefresh property on the article that
marks the article as dirty and would cause the article to be downloaded in the
background, but that is not the use case here)

Change-Id: I732a4ae14541032d7f9d05b5c93ad362900847cd