Fix for upstream css change affecting edit pencil.
The mobile app extension css changed a bit (an invisible border was added
around the edit pencils) which exposed a bug in the way the iOS app was
preventing collapsed tables from extending past the right edge of the
screen. This bug messed up the edit pencil layout when used in conjunction
with the upstream change to the css.

This patch scopes more narrowly the css preventing tables from
extending past the right edge of the screen, which fixes the edit pencil
layout issue.

Note: this is not needed for the 4.1.5 branch (unless we run
"make web" to pull the upstream css).

Bug: T102627
Change-Id: I048a272a48fe3bc110d630dd62495c25ff41b05a