make images scrollable in article header
Created a new, minimal gallery class specifically for the header:
WMFArticleHeaderGalleryViewController. It shares a parent class with the
original gallery: WMFPageCollectionViewController, which handles all
business logic for maintaining a "currentPage" property.

The article header gallery is embedded into the article VC's
tableheaderview. For MVP, the fullscreen gallery is presented modally
(as before) but we still hope to do a nicer transition in the near
future.

Also using a promise for the article request to allow presenting the
gallery while the mobileview data is still being loaded.

Also did various boyscout cleanups:
  - remove article loaded notification in fetcher & articleVC (using promise callbacks is
    sufficient)
  - refactored some UICollectionView utils (enumerating cells)
  - refactor RTL collection view logic into NSArray categories
  - refactor UICollectionViewLayout "fit item to collectionview bounds"
    logic into a custom layout—more reusable
  - fix gallery image owner label from continuing outside its cell

Added bonus! Modal image gallery now has loading indicators!

Bug: T102968
Change-Id: I166e1e3adabe8334a4fbeb06367fb0a7299ff37f