fix nil handling in langlink responses
When we get a response from langlinks where the langlinks field is
absent (instead of empty) for a given page, we need to gracefully return
an empty array.

See MWKLanguageLinkResponseSerializerTests for the example repsonse and
regression test.

This is a good example of why we should look into testing our fetchers
across at least a few hundred pages as part of a "slow" test suite.

Bug: T106853
Change-Id: I778fc261e1a77a92b653d90782e885c01e67d464