Removing ArticleFetcher
Replacing most logic with Serializers and WMFArticleFetcher

The big idea here is to start the ground work for https://phabricator.wikimedia.org/T106150
Will be adding a specific "ArticlePreview" object in the next few commits.

As noted previously, not all logic could be moved into the serializers.
Because of this, we have a base class WMFArticleBaseFetcher - which I loathe.
The purpose of it is to have a single place to keep the shared logic that cannot currently be placed in the serializers.