Fix underlying issue of the bounds error.
https://phabricator.wikimedia.org/T110381

The collection view layout is attempting to make calculations using Key Value Coding when the height of the cells is 0
(This occurs when there is a section, but not cells in the section)

This was still happening before the section was loaded when launching in landscape.
Launching in landscape causes the view to rotate prematurely kicking off layout logic.