Hide everything beneath lead image until preview data fetched.
Had to temporarily hide everything beneath the lead image until
the entire article is retrieved. This is because we've yet to
hook up/use the preview data we fetch, so it's hard to check for
it. When we do hook up the preview data we'll need to update
this check.