Fix sections crash
Checking isViewLoaded is not enough… apparently… to know whether it is safe to connect the collection view
You also have to check the window.
Not sure why the loading is a problem in the first place… but it is. And we don't have too much choice because the data source observes automatically