Removed networking from WebVC
Because the previous commit changed how the article was being set - I was left with 2 choices:
- Update all logic in the web view to work correctly with the new logic
- Remove any logic that is deprecated from the web view (fetching)

Since making the old code work with the new architecture would just be thrown away soon
and because it added a lot of complexity, this commit takes the latter approach.

This simplified the number of code paths to adapt to the new architecture.