make TOCVC use TOCItem protocol exclusively, refactor article container
To continue use Swift TableOfContentsItem protocol, we have to interact
w/ the TOC in Swift. This means that the portion of article container VC
code that uses the TOC also has to be in Swift, hence the Swift
extensions of article container VC and private header.  This is less
than elegant, but IMO I'd rather move _towards_ Swift than away from it.
The other option is to make TOCItem more ObjC-compatible (i.e. not have
Swift Enum properties), which would allow TOCVC to use it directly.