Fix bug where results are updated but there is lag before the delegate is notified.
Both of these controllers implicitly set their results when their underlying fetcher view model was updated.
Because of this there can be a time:
1. where a user scrolls causing the run loop to fire requesting a cell, and
2. The fetcher has completed its fetch and set the results to its own ivar, but
3. The promise with the results has yet to resolve.

This scenario causes the home VC to request an update to the cell thinking that the results are nil or placeholders, but the real results are available.
This didn't affect he app adversely until placeholder logic was added. In this scenario, the controller attempts to update a "real cell" with placeholder data since the placeholder data has yet to be replaced with the real data in the ssdatasource.