reconfigure all Wikipedia/Code files to be relative to proj
and localted in the "Code" folder.  Some groups had no "Location" in the
file attributes inspector.  This was resulting in file creation prompts
not starting in the expected directory, and who knows what else.

also moved Pods-acknowledgements into "Supporting Files" since it's not
code, and not relative to the "Code" folder