Make ReferencesContainerView modal for acessibility
Make assistive technologies (VoiceOver, Switch Control) treat
ReferencesContainerView as modal so that the text of the reference and
the text of the article do not mix together.

Also, when the References view is displayed, it is necessary to move the
accessibility cursor to this view, otherwise users with assistive
technologies cannot access its content.