Invoke js focus() for TOC only if VoiceOver is running and move javascript
This is a followup to "Move VoiceOver cursor to place in article on TOC
item selection" which makes sure the focus() method is only invoked in
the webView javascript when VoiceOver is running in order to prevent unwanted
scrolling for users who do not run assistive technologies.

This also moves the javascript code to www/js.