Work around for footer VC to fix visual glitch
The footer VC logic is not idempotent and cannot handle being refreshed.
This causes issue when refreshing the article.
This patch works around the imitation by not ever refreshing the footer since it is unlikely to change.
Filing a bug to fix this properly by fixing the footer logic to be able to be refreshed