Fix app shortcut showing black screen of death or doing nothing
When we moved the "app resume logic" to work on the didenterforeground notification, that inadvertently caused the "processshortcutitem" logic to run before the shortcut was ever delivered to the app.
This caused unpredictable behavior.

In the end, this is much cleaner and gets the shortcut processing out of the main resume logic - and removed 2 ivars…