Fix links not being open on cold start
There was a race condition… again… with the link coming through before the UI was created.

This time it was when we added the additional delay for making sure we didn't get unbalanced calls when showing the on boarding screen.

Now the method for processing these is resilient, and keeps the activity around until it can be processed.