Adding magic tap support
We can now close the table of contents in 2 ways:
1. Standard Escape gesture (WHICH HAS BEEN WORKING!!!!) - draw a "Z" with 2 fingers
2. Magic Tap - Double tap with 2 fingers

So the escape gesture should be known and should have been working. It is working for me now. Not sure if this ticket was still valid.
As a nicety I provided the magic tap gesture which is simpler as a close action as well.