Fix occasional flakiness positioning the bottom native container.
Problem is it looks like constraining to the bottom of the
WKContentView isn't reliable. Tried every which way - on the cases
where it's not positioning correctly - ie it thinks the bottom of
the page is actually somewhere above the bottom - nothing short of
reloading the entire page seems to correct it. I tried isolating
it and telling all the views they needed to update constraints and
layouts and forced layout updates but nothing seems to set it right.

Instead, I switched vertical constraint of the footer container to
the *top* of the scroll view (not the WKContentView!) and I
observe the scrollview's contentSize, which seems reliable, and
I adjust the constant on the top constraint to stay equal to the
contentSize height minus the footer container height.

With this approach no more WKWebView introspecting!!! :) Seems solid.
(note can't contrain to the bottom of the web view's scrollview
either - its height doesn't seem to get updated)