Fix test by adding Tulu language.
Note: Tulu is written in Kannada alphabet, but "kn" wiki is already
associated with "kn" localization. Since there is no Apple Tulu
"tcy" localization, we assume Tulu speakers are using "kn"
localization, but wanted to see Tulu as a search lang and in
the list of languages an article is available in. This PR adds
this.