Fix crash where mappings would crash because they were out of sync with the DB
The following:
NSArray *notifications = [self.articleReferenceReadConnection beginLongLivedReadTransaction];

Returns nil when out of process changes occur which usually means you have no work to do. But in this case we do.

Though Yap takes care to update the DB snapshot and all connections / extensions internally, it does not update any mappings in your app.

This means we need to propagate all notifications so mappings can be updated, even if it looks like there are no changes.