Fix 1st paragraph relocation to not be dependent on edit button element id.
Formerly it had depended on knowing the id of the first edit button. It would move the paragraph to just after that button. But now we can pass it the id of the element after which the paragraph will be moved. Solved an order-of-occurence bug preventing the paragraph and new edit transforms from working together.