Always allow tapping outside panel to dismiss it if it shows a close button.
Inverse is true as well - if no close button tapping outside the panel should not close it.

Chatted with Carolyn about this and it seemed safest option - especially since all of our panels so far have 1:1 correspondence between having a close button and whether tapping outside the panel dismisses it.