Use new pagelib reference collection bits.
This doesn't include the pagelib version bump because the PR moving those bits to the page lib isn't merged at the moment. As soon as https://github.com/wikimedia/wikimedia-page-library/pull/140 is merged we can bump the pagelib version (as a new commit on this PR) and it should work.