Add flag so we can tell if peeked item is a `Read more` item.
Adds a flag to the webVC which is used to record last time a `Read more` item was touched.

Then we simply can check in `webView:shouldPreviewElement:` if this happened recently - right now its checking if it happened within the last second.

This way was less invasive and didn't involve and pagelib or url hackery.