add placeholder label
UITextView doesn't have built-in placeholder, and we need to use UITextView vs UITextField for UITextView's ability to auto-grow vertically for long descriptions.
So we use a label for the placeholder, and we rely on stack view collapsing hidden subviews (which both the UITextView and the placeholder label are) when we toggle UITextView and the placeholder label `isHidden` values.
This approach also handles long placeholder text translations gracefully.