improve performance of select-to-edit
instead of potentially doing a series of regexes on the wikitext, do single regex and as we loop through each match grab a few words from either side of match and see how closely these match the few words we extracted from either side of the article text selection - a score is calculated to represent this closeness. highest scoring item is best candidate for wikitext match of the user's html selection.