Fix cross process notifier regression, fix missing explore feed cards
The unique channel name that needs to be the same in the app and the extension were no longer the same due to a bug in the way the keys were generated and persisted. The main context's channel name would be re-generated on every launch because only the cache context channel name was persisted.