Event Platform Client: Consolidate input and output buffers
Currently, the Event Platform Client persists events in two different
buffers: an "input buffer" used before stream configurations are loaded,
and an "output buffer" to which events are transferred from the input
buffer when stream configurations are loaded and which is used
thereafter.

This patch removes the input buffer as unnecessary and uses a single
event buffer throughout the app lifecycle. We would expect stream
configurations to be loaded very quickly on app startup in most cases.
In the event the periodic worker triggers event submission prior to
stream configurations being loaded, the submission is skipped and the
events remain in buffer.