Add and use user info to WMFPreferredLanguagesDidChangeNotification
Currently two MWKLanguageLinkController properties expose the
prior state of preferred languages to allow WMFAppViewController to
calculate if a reorder, add, or remove has happened and to which
language.

All of this info is available without calculation when the
notification is created. So, remove those methods, adds the
appropriate info to the user info and update notification method
to use the user info.

- Add two user info keys for WMFPreferredLanguagesDidChangeNotification
- Add enum of change types for WMFPreferredLanguagesDidChangeNotification
- Remove mostRecentlyModifiedPreferredLanguage property
- Remove previousPreferredLanguages
- Update WMFAppController method to use the user info data
- Remove unused testing methods / declarations

Change string constant name