Add MetricsClientBridge for Objective-C compatibility
Splits the EventPlatformClient's Objective-C compatilibity layer into
a bridging class which will remain in the app repository after the
client library is consolidated with the other platform libraries as
part of the work falling under T280182. This is needed because the
standalone Swift client library will need to run on Linux, where the
`@objc` annotation providing interoperability is not supported.

The bridging class will likely also be used to expose main app
functionality to the client library once the library itself is
extracted.

Note: Metrics Client is the new name for the Event Platform Client
project.