Let URLComponents setters handle encoding
A relative href can contain multiple components, such as the path,
query string, and fragment. Each component has a different set of
allowed characters that do not need to be encoded. We replace the
existing method of encoding relative hrefs with calls to the getters
and setters of URLComponents, which automatically decode and encode
based on the rules for each component.