In article share menu, only copy article URL
By default, the UIActivity.ActivityType.copyToPasteboard activity
concatenates all relevant items (such as strings and URLs) together and
puts the combined result into the pasteboard, which may not be what the
user expects.

We add two classes which can wrap around an item or a
UIActivityItemSource, and exclude a given list of activity types. When
an article's ShareActivityController is open, we exclude the
.copyToPasteboard activity from all items except for the article URL,
since the user likely expects the article URL to be the only thing
copied.