More progress towards images on talk pages
This commit adds a imple verification based on image height to determine if the image is an icon, and therefore should not be opened in a new view. Also changes the UIImage for icons. Adds other attributes to the image attributed strings. Currently the image indexes for images inside tables have issues with placement.