Table Views: Move to inset groups.
This change switches settings, libraries used, and the watchlist filtering view to
use the new inset grouped style that was introduced with iOS 14 and above. It also
migrates notification filtering and inbox views to use inset groups as well.  This is
a step towards modernizing the app to conform to modern iOS style guidelines.