Fix Phabricator links
The [ios-app-bugs](https://phabricator.wikimedia.org/tag/ios-app-bugs/) project in Phabricator was archived on 17 February 2023. Then, almost exactly 2 years later, on 12 February 2025, the Phabricator project [ios-bug-backlog](https://phabricator.wikimedia.org/tag/ios-bug-backlog/) was created, which seems to fulfil the exact same purpose.

So this patch replaces one with the other in the links, so new tasks won't get tagged with the archived project.