Remove providence tracker
PLEASE remove this junk. I know it only takes like five seconds
to remove it when I bookmark or share links but it's the five most
annoying seconds of every day—many times per day even since I link a
lot to Wikipedia. With this line removed, a lot of the other functions
in this file become nops so feel free to squash in some commits that
remove all of them as you merge this. Wikipedia is something we all
created for the public benefit, not to be annoying and painful like
commercial malware is.